move all export column tests to test_export_column
