+++ corehq/apps/export/tests/test_table_configuration.py
@@ -1,22 +1,15 @@
-from collections import namedtuple
-from mock import patch
 from django.test import SimpleTestCase
 
 from corehq.apps.export.const import USERNAME_TRANSFORM
-from corehq.apps.export.models import TableConfiguration, ExportColumn, \
-    ScalarItem, ExportRow
-from corehq.apps.export.models.new import (
-    SplitExportColumn,
-    MultipleChoiceItem,
-    Option,
+from corehq.apps.export.models import (
     DocRow,
     RowNumberColumn,
     PathNode,
-    StockExportColumn,
-    ExportItem,
+    ExportRow,
+    ScalarItem,
+    ExportColumn,
+    TableConfiguration,
 )
-
-MockLedgerValue = namedtuple('MockLedgerValue', ['product_id', 'section_id'])
 
 
 class TableConfigurationTest(SimpleTestCase):
@@ -311,102 +304,3 @@ class TableConfigurationGetRowsTest(SimpleTestCase):
                 ["0.1.1", 0, 1, 1, 'boop']
             ]
         )
-
-
-class SplitColumnTest(SimpleTestCase):
-
-    def test_get_value(self):
-        column = SplitExportColumn(
-            item=MultipleChoiceItem(
-                path=[PathNode(name='form'), PathNode(name='q1')],
-                options=[Option(value='a'), Option(value='c')]
-            ),
-            ignore_unspecified_options=False
-        )
-        doc = {"q1": "a b d"}
-        self.assertEqual(column.get_value(
-            'domain',
-            'docid',
-            doc,
-            [PathNode(name='form')],
-            split_column=True),
-            [1, None, "b d"]
-        )
-
-    def test_ignore_extas(self):
-        column = SplitExportColumn(
-            item=MultipleChoiceItem(
-                path=[PathNode(name='form'), PathNode(name='q1')],
-                options=[Option(value='a'), Option(value='c')]
-            ),
-            ignore_unspecified_options=True
-        )
-        doc = {"q1": "a b d"}
-        self.assertEqual(column.get_value(
-            'domain',
-            'docid',
-            doc,
-            [PathNode(name="form")],
-            split_column=True),
-            [1, None],
-        )
-
-    def test_basic_get_headers(self):
-        column = SplitExportColumn(
-            label="Fruit",
-            item=MultipleChoiceItem(
-                options=[Option(value='Apple'), Option(value='Banana')]
-            ),
-            ignore_unspecified_options=True
-        )
-        self.assertEqual(column.get_headers(split_column=True), ["Fruit | Apple", "Fruit | Banana"])
-
-    def test_get_headers_with_template_string(self):
-        column = SplitExportColumn(
-            label="Fruit - {option}",
-            item=MultipleChoiceItem(
-                options=[Option(value='Apple'), Option(value='Banana')]
-            ),
-            ignore_unspecified_options=True
-        )
-        self.assertEqual(column.get_headers(split_column=True), ["Fruit - Apple", "Fruit - Banana"])
-
-    def test_get_headers_with_extras(self):
-        column = SplitExportColumn(
-            label="Fruit - {option}",
-            item=MultipleChoiceItem(
-                options=[Option(value='Apple'), Option(value='Banana')]
-            ),
-            ignore_unspecified_options=False
-        )
-        self.assertEqual(
-            column.get_headers(split_column=True),
-            ["Fruit - Apple", "Fruit - Banana", "Fruit - extra"]
-        )
-
-
-class StockExportColumnTest(SimpleTestCase):
-    domain = 'stock-domain'
-
-    @patch('corehq.apps.export.models.new.StockExportColumn._get_product_name', return_value='water')
-    @patch('corehq.apps.export.models.new.Product.by_domain', return_value=[])
-    def test_get_headers(self, _, __):
-        column = StockExportColumn(
-            domain=self.domain,
-            label="Stock",
-            item=ExportItem(),
-        )
-        get_ledger_path = (
-            'corehq.form_processor.interfaces.dbaccessors.'
-            'LedgerAccessors.get_ledger_values_for_product_ids'
-        )
-        with patch(
-                get_ledger_path,
-                return_value=[
-                    MockLedgerValue(section_id='abc', product_id='def'),
-                    MockLedgerValue(section_id='abc', product_id='def'),
-                    MockLedgerValue(section_id='123', product_id='456'),
-                ]):
-
-            headers = list(column.get_headers())
-            self.assertEqual(headers, ['water (123)', 'water (abc)'])
